scm2_django_v2/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
channels>=4.0
daphne>=4.0
channels-redis>=4.1
psycopg2-binary>=2.9
django-filter>=23.0
Pillow>=10.0
celery>=5.3
redis>=5.0
